EWPP-1410: Rework entity type bundle names rendering.
